+++ webui/src/Buttons/ActionRecorder.jsx
@@ -35,7 +35,8 @@ import { useMemo } from 'react'
 import { DropdownInputField } from '../Components'
 import { ActionsList } from '../Controls/ActionSetEditor'
 import { GenericConfirmModal } from '../Components/GenericConfirmModal'
-import { ButtonGrid, ButtonGridHeader, usePagePicker } from './ButtonGrid'
+import { ButtonGrid, ButtonGridHeader } from './ButtonGrid'
+import { usePagePicker } from '../Hooks/usePagePicker'
 import { cloneDeep } from 'lodash-es'
 import jsonPatch from 'fast-json-patch'
 import { usePanelCollapseHelper } from '../Helpers/CollapseHelper'
